format, JavaDocs

+++ xchange-core/src/main/java/com/xeiam/xchange/utils/Base64.java
@@ -1028,7 +1028,7 @@ public class Base64 {
       } // end if: white space, equals sign or better
       else {
         // There's a bad input character in the Base64 stream.
-        throw new java.io.IOException(String.format("Bad Base64 input character decimal %d in array position %d", ((int) source[i]) & 0xFF, i));
+        throw new java.io.IOException(String.format("Bad Base64 input character decimal %d in array position %d", (source[i]) & 0xFF, i));
       } // end else:
     } // each input character
 
@@ -1083,7 +1083,7 @@ public class Base64 {
     boolean dontGunzip = (options & DONT_GUNZIP) != 0;
     if ((bytes != null) && (bytes.length >= 4) && (!dontGunzip)) {
 
-      int head = ((int) bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
+      int head = (bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
       if (java.util.zip.GZIPInputStream.GZIP_MAGIC == head) {
         java.io.ByteArrayInputStream bais = null;
         java.util.zip.GZIPInputStream gzis = null;
